Sends the expected DSCP codepoints
The SetTOS call makes a call to set the IP_TOS socket option (https://github.com/golang/net/blob/master/ipv4/sys_linux.go#L20), which sets the TOS field. The DSCP field occupies the higher 6 bits of this field, so values need to be right shifted by 2 in order to be the expected value.

Double check this with Wireshark because I haven't actually run the code, just spotted it.